backend/requirements.txt
tortoise-orm
asyncpg
uvicorn
fastapi
fastapi-users[tortoise-orm,oauth]
httpx_oauth
python-rapidjson
ciso8601
python-dotenv
passlib[bcrypt]
redis
requests
pyjwt
gunicorn
bitstring
oauth2client
google-auth
sendgrid